my eco > 내 글 조회 > 글 하나 클릭시 글 조회 activity이동

+++ eco100-app/app/src/main/java/com/mapo/eco100/entity/myeco/MyBoard.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 
 
 data class MyBoard(
-    val boardId: Int,
+    val boardId: Long,
     val nickname: String,
     val title: String,
     val date: String
